pyproject.toml
[project]
name = "gvgmiau"
version = "0.1.0"
description = ""
authors = [
    {name = "Martello"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-admin (>=2.0.2,<3.0.0)",
    "discord-py (>=2.6.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
